added stockdata viewmodel and added api calls in stock data repository

+++ app/src/main/java/com/example/stonks/data/StatisticsResponse.kt
@@ -22,7 +22,7 @@ data class SummaryDetail(
 )
 data class PreviousClose(
     @SerializedName("raw")
-    val value: Long
+    val value: Float
 )
 data class DefaultKeyStatistics(
     val sharesOutstanding: SharesOutstanding
